+++ libft/ft_strreplace.c
@@ -6,29 +6,11 @@
 /*   By: emadriga <emadriga@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/13 11:54:30 by emadriga          #+#    #+#             */
-/*   Updated: 2021/11/14 21:45:50 by emadriga         ###   ########.fr       */
+/*   Updated: 2022/01/05 19:50:28 by emadriga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
-
-static char	*ft_strjoin_freedouble(char *s1, char *s2)
-{
-	char	*join;
-	size_t	len;
-
-	if (s1 == NULL || s2 == NULL)
-		return (NULL);
-	len = ft_strlen(s1) + ft_strlen(s2) + 1;
-	join = ft_calloc(sizeof(char), len);
-	if (!join)
-		return (0);
-	ft_strlcat(join, s1, len);
-	ft_strlcat(join, s2, len);
-	free(s1);
-	free(s2);
-	return (join);
-}
 
 static char	*recursive_str_replace(const char *str, char const *oldset, \
 	char const *newset, size_t index)
@@ -53,7 +35,7 @@ static char	*recursive_str_replace(const char *str, char const *oldset, \
 }
 
 /**
- * * Returns a new string in which all occurrences of a specified string 
+ * * Returns a new string in which all occurrences of a specified string
  * * in the current instance are replaced with another specified string.
  * @param str		string to modify
  * @param oldset	string to be replaced.