+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: emadriga <emadriga@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/25 19:15:07 by emadriga          #+#    #+#             */
-/*   Updated: 2021/11/14 21:45:34 by emadriga         ###   ########.fr       */
+/*   Updated: 2022/01/05 19:49:24 by emadriga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -74,4 +74,5 @@ char	*ft_strreplace(const char *str, char const *oldset, char const *newset);
 char	*ft_strreplaceat(const char *str, char const *oldset, \
 		char const *newset, size_t index);
 int		ft_strcmp(const char *s1, const char *s2);
+char	*ft_strjoin_freedouble(char *s1, char *s2);
 #endif